Noise 3D
Applies noise based on the 3D position of image pixels. Noise3D is scaled according to the perspective, which is obviously more realistic when applying noise to an object or surface in an image. The noise can be output to any channel, but the alpha channel is the default.

Noise3D is in the 2.5D Toolbox, or find it in the TAB search tool as you can with all nodes.
Connect an AOV Bundle to the input, making sure the pass __Pworld has been generated for the CG