Create Ramp3D
Think of Ramp3D as an advanced ZSlice. It requires the presence of a Pworld pass with the pixel position data in the source image. The resulting ramp can be applied to any desired channel, though the default is the alpha channel.

Description of the node/tool
Ramp3D is in the 2.5D Toolbox, or find it in the TAB search tool as you can with all nodes.
Connect an AOV Bundle to the input, making sure the pass __Pworld has been generated for the CG



3D preview


You can compare the ramp to the actual depth of the pixels by opening the 3D preview tab and checking "active".
Then hover the mouse over the viewer and press the tab key to switch to 3D view. Press "f" or the middle mouse key to center the 3D image pixels in the viewer. The ramp is visible as a 3D shape extending between the start and end points. This allows you to see roughly where pixels fit on the ramp and adjust the ramp to encapsulate the desired range of the image. You can adjust the density of the pixel point cloud with the "quality" slider, and the size of the ramp object with the "ramp scale" slider.