Create Magic Pref
MagicPrefMap
Map 2D textures onto moving 3D renders using the Pref geometry pass

What does it do ?
MagicPrefMap creates a UV map that 'sticks' to a CG render, relative to a reference frame. This allows you to create a texture on one frame of animation then apply it to the sequence using an STMap node.

The plugin generates a point cloud map from the reference frame, which allows it to build a UV map for any frame of the animation. For more details see the How does it work ? section below.

MagicPrefMap uses the Pref AOV pass to generate the UV map. Unlike Pworld and Pobject, Pref values are fixed to the surface no matter what animation is applied to it, so a particular (x,y,z) value always maps to the same point on the geometry.

For a summary of the Pref and other standard AOVs please see the following link :

AOV Definitions old
How do I use it ?
The setup needed to map textures to CG renders is as follows :



Select a frame from the sequence to use as a reference
Create / Lineup the 2D texture over the CG element on the reference frame
The texture will appear to be projected onto the surface on the reference frame.
Create a MagicPrefMap node
For convenience, the reference frame will default to the current frame
Select the channels that contain the Pref data
Pref data should be unpremultiplied for correct mapping on edges
Including an alpha channel will be faster to process and take up less RAM as zero-alpha pixels will not be stored
Shuffling Pref into rgb and keeping the original alpha channel is a recommended workflow.
Connect the output of MagicPrefMap to the stmap input, and your 2D texture to the src input of an STMap node.
The UV map is in the red/green channels, normalised by default which is what STMap expects.
The output of the STMap will be the texture, deformed to the animating geometry.
Options
By default, the output of MagicPrefMap should 'just work' and produce a good quality UV map, but should you wish to play with the settings :

Ref Frame	<auto>	Reference frame where texture passes straight through without distortion. Defaults to the active frame when the node was created.
Pref channels	rgba	Pref source channels. Should be unpremultiplied
Skip	0	Skip scanlines while building the point cloud map. This will be faster to analyse and have smaller memory needs, at the cost of image quality
Normalise	True	Scale the map by image width / height.
Samples	8	How many points in the cloud to compare for each pixel. Loosely speaking more samples are better, and there's not much of a speed hit as you increase the value.
How does it work ?
MagicPrefMap builds a point cloud from a CG render's Pref geometry pass. Each point in the cloud stores the (x,y) coordinate of the source pixel on the reference frame that was analysed.

Generating the map on any other frame is simply a case of finding the closest points in the cloud analysed before that match the current pixels Pref, and copying the pixel coordinates stored there into the output. Because Pref is fixed on the surface of the geometry, you now have a map that sticks to the surface of the object.

The skip knob reduces the size of the point cloud map, by skipping scanlines on the reference frame. This is a relatively crude way to reduce the size of the map which will affect quality of the UV map, although increasing samples afterwards can improve the result.

Tips / Known bugs / Future work
This is the first version of this tool. If you find any bugs or have any suggestions for features please log a support ticket.

Tips

If you can, choose a reference frame where the subject is closest to the camera.
This will give you a high quality point cloud map, and hence a good UV map
If you do not provide an alpha channel, every pixel in the image will be added to the point cloud map. 
This will greatly increase the size of the point cloud map, for no apparent increase in image quality

Known bugs

The edges of the bbox are not 'black outside' so you may see colours bleeding around the edge of the frame.
Including an alpha channel with the texture map will allow you to control this - just unpremultiply the input to the STMap, and premultiply the output